+++ packages/migrate/CHANGELOG.md
@@ -1,5 +1,27 @@
 # svelte-migrate
 
+## 1.0.0-next.4
+
+### Patch Changes
+
+- handle lone return statements ([#5831](https://github.com/sveltejs/kit/pull/5831))
+
+* Fix error placement on (arrow) function when checking load input ([#5840](https://github.com/sveltejs/kit/pull/5840))
+
+## 1.0.0-next.3
+
+### Patch Changes
+
+- handle more import cases ([#5828](https://github.com/sveltejs/kit/pull/5828))
+
+* check load function input ([#5838](https://github.com/sveltejs/kit/pull/5838))
+
+## 1.0.0-next.2
+
+### Patch Changes
+
+- Correctly rename files with spaces when migrating ([#5820](https://github.com/sveltejs/kit/pull/5820))
+
 ## 1.0.0-next.1
 
 ### Patch Changes